facturation goof

+++ back-end/database/migrations/2024_04_02_195036_create_ligne_cartes_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
 
         Schema::dropIfExists('lignecarte');
         Schema::create('lignecarte', function (Blueprint $table) {
-            $table->bigIncrements('id');  // Use bigIncrements() for auto-incrementing big integer primary key
+            $table->increments('id');  // Use bigIncrements() for auto-incrementing big integer primary key
             $table->unsignedInteger('idFac');
             $table->unsignedInteger('idCarte');
             $table->integer('point');
